Back to float for testing

+++ reachables/src/main/java/com/act/lcms/v2/MassToRawMetaboliteMap.java
@@ -11,7 +11,7 @@ import java.util.NavigableMap;
 import java.util.TreeMap;
 
 /**
- * This class holds the API for a {Double mass -> List<RawMetabolite> rawMetabolites} map.
+ * This class holds the API for a {Float mass -> List<RawMetabolite> rawMetabolites} map.
  * Its purpose is to store in memory enumerated lists of structures or formulae and allow for quick lookups.
  * Along with the NavigableMap holding the data structure, an enum defines the expected kind of metabolite.
  * Note: the expectation is that the conversion to actual Metabolite objects is done outside of this API.
@@ -23,7 +23,7 @@ public class MassToRawMetaboliteMap {
 
   // The data structure holding the RawMetabolites.
   // A NaviagableMap has very convenient properties (sorted, API for extracting sub-maps) for our use case.
-  private NavigableMap<Double, List<RawMetabolite>> massToRawMetaboliteMap;
+  private NavigableMap<Float, List<RawMetabolite>> massToRawMetaboliteMap;
   private RawMetaboliteKind kind;
 
   // Sub-class defining the kind of metabolite. Defined at the top-level to avoid storing it in every RawMetabolite.
@@ -36,7 +36,7 @@ public class MassToRawMetaboliteMap {
     this.kind = kind;
   }
 
-  public NavigableMap<Double, List<RawMetabolite>> getMassToMoleculeMap() {
+  public NavigableMap<Float, List<RawMetabolite>> getMassToMoleculeMap() {
     return massToRawMetaboliteMap;
   }
 
@@ -47,7 +47,7 @@ public class MassToRawMetaboliteMap {
   /**
    * Add a RawMetabolite to the map
    */
-  public void add(Double mass, RawMetabolite rawMetabolite) {
+  public void add(Float mass, RawMetabolite rawMetabolite) {
     List<RawMetabolite> matchingMetabolites = massToRawMetaboliteMap.get(mass);
     if (matchingMetabolites == null) {
       matchingMetabolites = new ArrayList<>();
@@ -59,14 +59,14 @@ public class MassToRawMetaboliteMap {
   /**
    * Retrieve all the RawMetabolites within a given mono-isotopic mass window
    */
-  public Map<Double, List<RawMetabolite>> getMassWindow(Double minMass, Double maxMass) {
+  public Map<Float, List<RawMetabolite>> getMassWindow(Float minMass, Float maxMass) {
     return massToRawMetaboliteMap.subMap(minMass, maxMass);
   }
 
   /**
    * Retrieve all the RawMetabolites within a given centered mono-isotopic mass window
    */
-  public Map<Double, List<RawMetabolite>> getMassCenteredWindow(Double center, Double windowSize) {
+  public Map<Float, List<RawMetabolite>> getMassCenteredWindow(Float center, Float windowSize) {
     return getMassWindow(center - windowSize / 2, center + windowSize / 2);
   }
 
@@ -74,9 +74,9 @@ public class MassToRawMetaboliteMap {
    * Retrieve a sorted list of the RawMetabolites within a given a centered mono-isotopic mass window,
    * ordered by their closeness to the center.
    */
-  public List<RawMetabolite> getSortedFromCenter(Double center, Double windowSize) {
-    Map<Double, List<RawMetabolite>> subMap =  getMassCenteredWindow(center, windowSize);
-    Map<Double, List<RawMetabolite>> newMap = new TreeMap<>();
+  public List<RawMetabolite> getSortedFromCenter(Float center, Float windowSize) {
+    Map<Float, List<RawMetabolite>> subMap =  getMassCenteredWindow(center, windowSize);
+    Map<Float, List<RawMetabolite>> newMap = new TreeMap<>();
     subMap.entrySet().stream().forEachOrdered(entry -> newMap.put(Math.abs(entry.getKey() - center), entry.getValue()));
     List<RawMetabolite> rawMetabolites = new ArrayList<>();
     newMap.values().forEach(rawMetabolites::addAll);
